portfolio-backend-public/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
litellm==1.70.0
MarkupSafe==3.0.2
multidict==6.4.3
openai==1.75.0
packaging==25.0
pluggy==1.6.0
prometheus_client==0.22.1
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.13.0
pytest==8.2.1
pytest-asyncio==0.23.6
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.0
sniffio==1.3.1
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
zipp==3.21.0
